requirements/dev.txt
ipython==8.7.0
coverage
django-debug-toolbar